requirements.txt
blinker==1.9.0
click==8.3.0
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Faker==25.2.0
filelock==3.20.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.60.1
fsspec==2025.10.0
greenlet==3.2.4
ImageIO==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.2
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
lime==0.2.0.1
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==1.26.2
packaging==25.0
pandas==2.1.3
pillow==12.0.0
psycopg2-binary==2.9.9
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
scikit-image==0.25.2
scikit-learn==1.3.2
scipy==1.16.3
shap==0.49.1
six==1.17.0
slicer==0.0.8
SQLAlchemy==2.0.23
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.10.16
torch==2.9.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.0.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-huggingface==0.1.2
langchain-chroma==0.1.4
chromadb==0.5.23
sentence-transformers==3.4.1
groq==0.18.0
gunicorn==22.0.0

